+++ src/components/DataCard.js
@@ -2,37 +2,13 @@ import React from 'react';
 import './DataCard.css';
 import useFetch from '../hooks/useFetch';
 import OneCard from './OneCard';
-
-function MultiCard({ isLoading, data, keyPrefix, keySuffixPrepend }) {
-  const _data = data ?? [{ continent: 'continent' }];
-
-  const components = _data.map(item => (
-    <>
-      <header>
-        <h3 className="article-title">{item.continent.toUpperCase()}</h3>
-      </header>
-      <OneCard
-        data={item}
-        isLoading={isLoading}
-        keyPrefix={keyPrefix}
-        keySuffixPrepend={`${keySuffixPrepend}-${item.continent}`}
-      />
-    </>
-  ));
-
-  return <>{components}</>;
-}
+import MultiCard from './MultiCard';
 
 function DataCard({ url, params, title, keyPrefix, multi = false }) {
-  const {
-    data,
-    isLoading,
-    hasError,
-    errorMessage,
-    updateUrl,
-    updateParams,
-    refetch,
-  } = useFetch(url, params);
+  const { data, isLoading, hasError, errorMessage, refetch } = useFetch(
+    url,
+    params
+  );
 
   if (hasError) return <div>{errorMessage}</div>;
 
@@ -54,7 +30,7 @@ function DataCard({ url, params, title, keyPrefix, multi = false }) {
 
   return (
     <article className="DataCard">
-      <header>
+      <header className="article-header">
         <h3 className="article-title">{title}</h3>
         <button onClick={refetch}>refresh</button>
       </header>
