+++ src/components/MultiCard.css
@@ -1,15 +1,15 @@
-.DataCard {
+.MultiCard {
   color: var(--white);
   background-color: var(--brown-grey);
 }
 
-.DataCard > .article-header {
+.MultiCard > .article-header {
   padding: 8px 0;
   display: flex;
   justify-content: space-around;
   align-items: center;
 }
-.DataCard > footer {
+.MultiCard > footer {
   padding: 8px 0;
   display: flex;
   justify-content: space-around;
